refactor: separate types in their own file and the test in own file
